+++ changelog.md
@@ -5,12 +5,15 @@
 ##### Added
 
 - Markdown file top toolbar added
+- Added `?theme=` support loading custom css
+- Added ~~strike~~ support
 
 ##### Fixed
 
 - Fixed inconsistency with markdowns in files
 - Syntax was reading whilest in code blocks
 - Fixed layered script
+- Changelog issues fix
 
 #### Version 1.0
 
@@ -18,8 +21,6 @@
 
 - Initial website preparation
 - GitHub markdown implementation
-- Setup default website state
 - File not found fail state prepared
-- Added Customization/Settings window
 
 > It's time to make something burn